refactor(sdk): Change waveform to be a list of values between 0 and 1
Most clients will probably work with values between 0 and 1 and need to
convert it just to send it, so we can move that conversion into the SDK.

This is also more forwards-compatible, because MSC3246 now has a
different max value for the amplitude, so when this becomes stable, the
only change needed will be in the SDK.

+++ bindings/matrix-sdk-ffi/src/timeline/mod.rs
@@ -424,7 +424,7 @@ impl Timeline {
         self: Arc<Self>,
         params: UploadParameters,
         audio_info: AudioInfo,
-        waveform: Vec<u16>,
+        waveform: Vec<f32>,
     ) -> Result<Arc<SendAttachmentJoinHandle>, RoomError> {
         let mut info =
             BaseAudioInfo::try_from(&audio_info).map_err(|_| RoomError::InvalidAttachmentData)?;
